+++ ui/src/components/index.ts
@@ -5,3 +5,4 @@ export * from './Interpret';
 export * from './Interpreters';
 export * from './TaskCards';
 export * from './Attack';
+export * from './Attacks';
